+++ sprout/weed/parser/numeric/int.hpp
@@ -1,6 +1,7 @@
 #ifndef SPROUT_WEED_PARSER_NUMERIC_INT_HPP
 #define SPROUT_WEED_PARSER_NUMERIC_INT_HPP
 
+#include <cstddef>
 #include <cstdint>
 #include <sprout/config.hpp>
 #include <sprout/integer/integer_digits.hpp>
@@ -11,7 +12,7 @@ namespace sprout {
 		//
 		// int_
 		//
-		SPROUT_STATIC_CONSTEXPR auto int_ = sprout::weed::int_p<std::intmax_t, 10, 1, -1>();
+		SPROUT_STATIC_CONSTEXPR auto int_ = sprout::weed::int_p<std::intmax_t, 10, 1, static_cast<std::size_t>(-1)>();
 
 		//
 		// int8
